pyproject.toml
[project]
name = "django-serializers"
version = "0.1.0"
description = ""
authors = [
    {name = "RNT13",email = "renatornt13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "django_serializers", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]
